Update base.html.twig handling favicon image
You create a new folder in the `/user/pages/images` if it is not already there and stick all of your favicons there, you can use any file format the browser accepts, including .ico (usually you use .ico or .png for favicon images) Then if you want to use a custom favicon, on the frontmatter of the page you put this parameter
`custom_favicon: whatever.ico`
If the `custom_favicon` header variable is not there, it will use the default favicon.ico.

You can use `theme://images/` if you prefer to keep everything in the theme images folder, but it is subject to rewriting if you update theme.  User folder or even individual pages folders are more permanent for custom content.